+++ components/core/ui/theme/src/main/kotlin/com/flipperdevices/core/ui/theme/composable/FlipperPalletProvider.kt
@@ -1,5 +1,7 @@
 package com.flipperdevices.core.ui.theme.composable
 
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -21,15 +23,36 @@ fun getThemedFlipperPallet(systemIsDark: Boolean): FlipperPallet {
 }
 
 @Composable
+fun getThemedFlipperPallet(theme: SelectedTheme): FlipperPallet {
+    return when (theme) {
+        // TODO for new theme
+        SelectedTheme.LIGHT -> lightPallet
+        SelectedTheme.DARK -> darkPallet
+        else -> getThemedFlipperPallet(systemIsDark = isSystemInDarkTheme())
+    }
+}
+
+@Composable
 fun isLight(systemIsDark: Boolean): Boolean {
     val themeViewModel: ThemeViewModel = viewModel()
     val theme by themeViewModel.getAppTheme().collectAsState()
 
     return when (theme) {
+        // TODO for new theme
         SelectedTheme.LIGHT -> true
         SelectedTheme.DARK -> false
         else -> !systemIsDark
     }
+}
+
+fun setAppCompatDelegateTheme(theme: SelectedTheme) {
+    val systemThemeId = when (theme) {
+        // TODO for new theme
+        SelectedTheme.LIGHT -> AppCompatDelegate.MODE_NIGHT_NO
+        SelectedTheme.DARK -> AppCompatDelegate.MODE_NIGHT_YES
+        else -> AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM
+    }
+    AppCompatDelegate.setDefaultNightMode(systemThemeId)
 }
 
 @Suppress("MagicNumber")