+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [Feature] Add save as on nfc editor
 - [Feature] Add dialog on exit from nfc editor
 - [REFACTOR] Migrate synchronization task graph to TaskSynchronizationComponent
+- [FIX] Choose app theme from selected theme, not is light
 
 # 1.2.1